# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Calendar API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /items:
    get:
      summary: List items
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: kind, schema: { enum: [event, task, both] } }
        - { in: query, name: status, schema: { enum: [active, partial, completed, cancelled, deleted] } }
      responses:
        "200": { description: Item list }
    post:
      summary: Create an item
      parameters:
        - { in: query, name: force, schema: { type: boolean } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ItemInput" } } }
      responses:
        "201": { description: Created }
        "409": { description: Conflict confirmation required }
...
components:
  securitySchemes:
    sessionCookie:
    demoCookie:
      type: apiKey
      in: cookie
      name: calendar_demo_session

      type: apiKey
      in: cookie
      name: better-auth.session_token
  schemas:
    ItemInput:
      type: object
      required: [title, kind]
      properties:
        id: { type: string, format: uuid }
        kind: { enum: [event, task, both] }
        title: { type: string }
        description: { type: string }
        location: { type: string }
        startAt: { type: [string, "null"], format: date-time }
        endAt: { type: [string, "null"], format: date-time }
        dueAt: { type: [string, "null"], format: date-time }
        isAllDay: { type: boolean }
        timezone: { type: string }
        priority: { enum: [none, low, medium, high, urgent] }
        status: { enum: [active, partial, completed, cancelled, deleted] }
        autoRollover: { type: boolean }
        showInTimetable: { type: boolean }
        parentId: { type: [string, "null"], format: uuid }
        recurrence: { type: [object, "null"], properties: { weekParity: { enum: [all, odd, even] } } }
        tagIds:
          type: array
          items: { type: string, format: uuid }
        reminders:
          type: array
          items: { type: object, properties: { trigger: { enum: [before_start, at_start, at_end] }, offsetMinutes: { type: integer }, channels: { type: array, items: { enum: [in_app, browser_push, email] } } } }